Rename the class to `ConsulProviderFactoryTests`. And file name should match class name.
